src/main/java/com/ayo/metric/services/MetricConversionFactory.java
package com.ayo.metric.services;

import org.springframework.http.HttpStatus;
import org.springframework.stereotype.Service;
import org.springframework.web.server.ResponseStatusException;

@Service
public class MetricConversionFactory {
    public MetricConversion getConversion(String type) {
		if(type.equalsIgnoreCase("TEMPERATURE")) {
			return new TemperatureConversion();
		} else if (type.equalsIgnoreCase("LENGTH")) {
			return new LengthConversion();
		} else if (type.equalsIgnoreCase("AREA")) {
			return new AreaConversion();
		} else if (type.equalsIgnoreCase("VOLUME")) {
			return new VolumeConversion();
		} else if (type.equalsIgnoreCase("WEIGHT")) {
			return new WeightConversion();
		} else {
			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Invalid Category of Conversion <" + type + ">");
		}
	}
}
